Move book data to Ultima\Data\Books.cs
